Improve the parallelism of random search
This breaks the search up into chunks, sending collections of solutions to the inspector as a group. This reduces the amount of contention by reducing the number of times we have to unlock the Mutex.

We also had to then send arrays of a single solution to the inspector in the serial version of random search.