Make `update_best` and `print_best` generic
This gets these to match by making both functions generic. I suspect that what we really want is an `Inspector` type with these as methods in that type. That would avoid having to declare and manage `best` in the `main()` loop, and passing `best` in to `update_best`, which all sounds nice.