Finish up sequential hill-climber
This gets the sequential hill-climber working in a pretty reasonable way. Lots of help from esitsu@Twitch and @JustusFluegel on how to improve the structure of my initial implementation.

We spent a _lot_ of time flailing around attempts to use `max_by_key`, which just created all kinds of ownership and lifetime issues. Ultimately we just went with `max_by`, which works fine, but certainly isn't as nice as `max_by_key` would have been if we'd been able to get that to work.

Big thanks to esitsu@Twitch for the suggestion to use `slice::from_ref` to turn `&current_scored_best` into a `&[current_scored_best]` without copying. This really cleaned up the call to the `inspector` and allowed me to unwind some icky business that I had added to deal with references.