Switch from lexicase to tournament selection
This is to match DEAP, although this uses tournament size 10 instead of 3 because 10 gave me better success rates. I think that if we had elitism, we would also get better success rates without having to increase the tournament size, but I'm not 100% sure.

I do know that we got better success rates with lexicase selection.