Deal with simplifying an empty genome
If none of the instructions in the `Plushy` genome actually _do_ anything, then simplification can get down to an empty genome and then panic because there's not an instruction to remove.

I've dealt with this by checking for an empty genome in the simplification loop, but we should probably also address this in `drop_random_instruction` as well.

I also added some comments about possibly moving some of the logic to `Plushy` and `TestResults`.