Add `complex_regression` implementation
This finds solutions fairly often. I had to bump up the population size to 1K and run for 100 generations, and I needed `FloatInstruction::Dup`, but with those things, I get fairly plausible results.

I had to bump up the `PENALTY_VALUE` to 1M so that the system didn't "decide" that "doing nothing" was better than "trying.