Fix error introduced by `abs_sub()`
It was suggested on a recent stream that I could replace `(x - y).abs()` with `abs_sub(x, y)`, and I went along with it. It turns out that you _can't_ do that, because `abs_sub()` assumes that `x>y`, which we do _not_ know is true in this instance.

That took me a while to debug, but I think it's fixed now.